+++ src/models/availability.rs
@@ -18,6 +18,7 @@ pub enum AvailabilityStatus {
 #[serde(rename_all = "snake_case")]
 pub enum AvailabilitySource {
     RouteRejected,
+    RouteUnverified,
     HarnessInstalled,
     UniversalHarness,
     #[serde(rename = "pi_probe")]